NI-460 - add conditional styling (#56)
