+++ Sources/RoktUXHelper/Services/LayoutTransformer/LayoutTransformer.swift
@@ -419,10 +419,19 @@ where CreativeSyntaxMapper.Context == CreativeContext {
                 accessibilityGrouped: Bool = false) throws -> RowViewModel {
         let updatedStyles = try StyleTransformer.updatedStyles(styles?.elements?.own, transform: BaseStyles.init)
 
-        return RowViewModel(children: children,
-                            stylingProperties: updatedStyles,
-                            accessibilityGrouped: accessibilityGrouped,
-                            layoutState: layoutState)
+        return RowViewModel(
+            children: children,
+            stylingProperties: updatedStyles,
+            animatableStyle: AnimationStyle(
+                transition: styles?.conditionalTransitions,
+                transform: { $0.own.map(BaseStyles.init) }
+            ),
+            accessibilityGrouped: accessibilityGrouped,
+            layoutState: layoutState,
+            predicates: styles?.conditionalTransitions?.predicates,
+            globalBreakPoints: layoutPlugin.breakpoints,
+            offers: layoutPlugin.slots.map(\.offer)
+        )
     }
 
     func getScrollableRow(_ styles: LayoutStyle<ScrollableRowElements,
@@ -431,10 +440,19 @@ where CreativeSyntaxMapper.Context == CreativeContext {
                           accessibilityGrouped: Bool = false) throws -> RowViewModel {
         let updatedStyles = try StyleTransformer.updatedStyles(styles?.elements?.own, transform: BaseStyles.init)
 
-        return RowViewModel(children: children,
-                            stylingProperties: updatedStyles,
-                            accessibilityGrouped: accessibilityGrouped,
-                            layoutState: layoutState)
+        return RowViewModel(
+            children: children,
+            stylingProperties: updatedStyles,
+            animatableStyle: AnimationStyle(
+                transition: styles?.conditionalTransitions,
+                transform: { $0.own.map(BaseStyles.init) }
+            ),
+            accessibilityGrouped: accessibilityGrouped,
+            layoutState: layoutState,
+            predicates: styles?.conditionalTransitions?.predicates,
+            globalBreakPoints: layoutPlugin.breakpoints,
+            offers: layoutPlugin.slots.map(\.offer)
+        )
     }
 
     func getZStack(_ styles: LayoutStyle<ZStackElements, ConditionalStyleTransition<ZStackTransitions, WhenPredicate>>?,
