+++ Sources/RoktUXHelper/UI/Components/ViewModel/RowViewModel.swift
@@ -11,14 +11,24 @@
 
 import Foundation
 import DcuiSchema
+import Combine
 
 @available(iOS 15, *)
-class RowViewModel: Identifiable, Hashable, ScreenSizeAdaptive {
+class RowViewModel: Identifiable, Hashable, ScreenSizeAdaptive, AnimatableStyleHandling {
     let id: UUID = UUID()
     var children: [LayoutSchemaViewModel]?
     let stylingProperties: [BasicStateStylingBlock<BaseStyles>]?
     let accessibilityGrouped: Bool
     weak var layoutState: (any LayoutStateRepresenting)?
+    let animatableStyle: AnimationStyle?
+    let predicates: [WhenPredicate]?
+    let globalBreakPoints: BreakPoint?
+    let offers: [OfferModel?]
+    var width: CGFloat = 0
+    var cancellable: AnyCancellable?
+
+    @Published var animate: Bool = false
+
     var imageLoader: RoktUXImageLoader? {
         layoutState?.imageLoader
     }
@@ -29,11 +39,22 @@ class RowViewModel: Identifiable, Hashable, ScreenSizeAdaptive {
 
     init(children: [LayoutSchemaViewModel]?,
          stylingProperties: [BasicStateStylingBlock<BaseStyles>]?,
+         animatableStyle: AnimationStyle?,
          accessibilityGrouped: Bool,
-         layoutState: (any LayoutStateRepresenting)?) {
+         layoutState: (any LayoutStateRepresenting)?,
+         predicates: [WhenPredicate]?,
+         globalBreakPoints: BreakPoint?,
+         offers: [OfferModel?]) {
         self.children = children
         self.stylingProperties = stylingProperties
+        self.animatableStyle = animatableStyle
         self.accessibilityGrouped = accessibilityGrouped
         self.layoutState = layoutState
+        self.predicates = predicates
+        self.globalBreakPoints = globalBreakPoints
+        self.offers = offers
+
+        animate = shouldApply(width)
+        subscribeToAnimation()
     }
 }
